move json list operations to separate trait
